backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.8
python-dotenv==1.0.0
httpx>=0.24.0,<0.25.0
pytest==7.4.3
pytest-asyncio==0.21.1
supabase==2.0.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
resend==0.6.0 